[refactor] react-query > todo 관련 폴더화
