+++ client/pages/[id].tsx
@@ -4,8 +4,8 @@ import TodoList from '../components/TodoList';
 import CreateTodoForm from '../components/CreateTodoForm';
 import {useRecoilState} from 'recoil';
 import {detailStateAtom, POST_STATE} from '../recoil/atoms/state';
-import {useDelete} from '../react-query/useDelete';
-import useGet from '../react-query/useGet';
+import {useDelete} from '../react-query/todos/useDelete';
+import useGet from '../react-query/todos/useGet';
 import EditTodoForm from '../components/EditTodoForm';
 import Header from '../components/Header';
 import Button from '../atomic/Button';
