+++ client/react-query/todos/useDelete.ts
@@ -1,4 +1,4 @@
-import todo from '../pages/api/todo';
+import todo from '../../pages/api/todo';
 import {useMutation, useQueryClient} from "@tanstack/react-query";
 import {todoKeys} from "./index";
