+++ client/react-query/todos/useGetAll.ts
@@ -1,4 +1,4 @@
-import Todo, { todo } from '../pages/api/todo';
+import Todo, { todo } from '../../pages/api/todo';
 import {useQuery} from "@tanstack/react-query";
 import {todoKeys} from "./index";
