+++ client/react-query/todos/useUpdate.ts
@@ -1,5 +1,6 @@
-import Todo from '../pages/api/todo';
+import Todo from '../../pages/api/todo';
 import {useMutation, useQueryClient} from "@tanstack/react-query";
+import {todoKeys} from "./index";
 
 interface UseUpdate {
   id: string;
